+++ __test__/domutil.test.js
@@ -1,5 +1,4 @@
-import domutil from '@/lib/domutil';
-import helper from '@/lib/helper';
+import * as domutil from '@/lib/domutil';
 
 const eventType = 'click';
 let mockFn = jest.fn();
@@ -19,44 +18,6 @@ beforeAll(() => {
 afterEach(() => {
   mockFn.mockReset();
   domutil.removeAll(element, eventType);
-});
-
-describe('요소 접근', () => {
-  test('selector로 요소 접근하는 경우', () => {
-    // given
-    const selectorId = element.id;
-
-    // when
-    const received = helper.getElement(`#${selectorId}`);
-
-    // then
-    expect(received).toEqual(element);
-  });
-
-  test('DOM Element로 접근 하는 경우', () => {
-    // given
-    const selectorDOM = element;
-
-    // when
-    const received = helper.getElement(selectorDOM);
-
-    // then
-    expect(received).toEqual(element);
-  });
-
-  test('유효하지 않은 값으로 요소 접근하는 경우', () => {
-    // given
-    const selectorEmptyString = '';
-    const selectorNull = null;
-
-    // when
-    const throwErrorFn1 = () => helper.getElement(selectorEmptyString);
-    const throwErrorFn2 = () => helper.getElement(selectorNull);
-
-    // then
-    expect(throwErrorFn1).toThrow();
-    expect(throwErrorFn2).toThrow();
-  });
 });
 
 describe('이벤트 바인딩 (공통)', () => {
